Refactored much of the action in PGProblemEditor.pm  This allows a user inteface which mimics the interface for the Hmwk sets editor and for the Classlist editor.
It also allows all of the temporary edit files to be placed in one location.
This allows one to "temporarily" edit a library file to try to see what is wrong with it.

It is also now much easier to make a local copy of a library file for your current problem set
so that you can tweak it.

It has been fairly well tested by one person (me) but there may still be bugs.

-- Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3718 c0722133-6baf-4dd8-8699-98d999cd4f06