Improved handling of case where a blank_problem is asked for in PGProblemEditor.pm Also improved the random names for new local problems
Added ability to create a new blank problem in ProblemSetDetail.pm (without going to the PGProblemEditor.pm)
however something is not yet working when this problem record is saved.

 Sam -- could you look at
lines near 679 to see what the problem is?  There is a FIXME note there.   I suspect that the addProblemToSet
routine is failing somehow, but there is no error message.  Many of the record fields are not filled out --
including the sourceFile path and the number of attempts.  The same snippet of code works fine
in PGProblemEditor.pm

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3719 c0722133-6baf-4dd8-8699-98d999cd4f06