Modifications that set problems with path .../Library/setFoo/problemBar to reference the "Library" component in bugzilla.
Also the problemSeed is added to the url sent to bugzilla e.g.

/webwork/shared-courses/gage_course_single_sql/templates/rochesterLibrary/setMAAtutorial/simplemultiplechoiceexample.pg_with_problemSeed=4655

this means that you can't click on the link, but it automatically includes a seed which might be helpful.
It is not guaranteed that this the seed that caused problems -- that depends on how the
bug is submitted -- but it is often the case that this seed caused the problem.

-- Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3779 c0722133-6baf-4dd8-8699-98d999cd4f06