Turn off debugging.
Allow the theme to be edited for each course. (This is slightly dangerous
since a misspelling of the theme could cause troubles.)
It would be helpful to have a pop-up menu widget for Config.pm -- similar to
the boolean or permission pop ups, but where the values
can be specified.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3859 c0722133-6baf-4dd8-8699-98d999cd4f06