Resolves bug #942, in which a missing permission level field causes a warning:
This is in fact due to UserList assuming that $record{permission} is
defined when it is allowed to be undefined (as per the format of
classlist files).

I added a case that sets is to $default_permission_level if it is not
defined.

FIXED in branch rel-2-2-dev revision 1.75.2.1.
Forwarded-ported with no modifications to HEAD in this revision.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3894 c0722133-6baf-4dd8-8699-98d999cd4f06