forward port: (fix width issues with info box.
restrict width of info box to exactly 40% of the page width. we can't
use "max-width" because MSIE doesn't support it.

set "overflow:auto" to enable scroll bars when the content can't be
wrapped to 40%. this is a little nicer than the contents running over
the border of the box, but MSIE makes bad choices about how to size the
box so as to avoid unnecessary scroll bars. (it sizes the element first
and then adds scroll bars, so it adding a horizontal scroll makes the
content area smaller than the content, necessitating a vertical scroll
bar as well. pretty annoying.)

Tested this in WinIE, Firefox, and Safari.)


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@3949 c0722133-6baf-4dd8-8699-98d999cd4f06