Implemented unarchiving function to complement archiving function.
Archive exports the course database to courseName/DATA and then tas and gzips the directory courseName to
courseDirectory/courseName.tar.gz

Unarchive allows one to choose files of the type courseDirectory/courseName.tar.gz  and then
ungzip, untar them to courseDirectory/courseName   -- the database record in courseName/DATA is
then loaded into the webwork database.

So far there is only minimal error checking about whether all or part of the course exists when unarchiving.

There are not yet features for archiving and then deleting a course -- these two operations have to be done
separately for now.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4129 c0722133-6baf-4dd8-8699-98d999cd4f06