Changed messages that refer to 'Save as' to use 'Create a copy' instead (which is the current wording for that function).
Problem:  Since the default in that pop-up menu us "Rename file"
rather than "Create a copy", it is not immediately clear what this
message refers to.

I still think "Create a copy" should be the default rather than
"Rename file", since the latter only makes sense when you got to the
editor from a file linked to a homework set, which is not always the
case (e.g., when you use the Library Browser).  It would also make
these error messages refer to something visible rather than a menu
item that is not showing.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4164 c0722133-6baf-4dd8-8699-98d999cd4f06