Fixed a couple more issues with blank problems.  Allow files to be copied even when we can't write the original (this lets you make a copy of the blank problem or default headers).
Attempt to overcome the copy/rename problem by making it not be a menu
and instead use a checkbox that says whether the problem or header
will replace the one in the current set.  This makes it clear that a
copy is still being made (not that the original is being renamed), but
that the homework set is being altered.  Those are really two
different actions, and it seems to make sense to separate them.  The
default is still to link into the homework set, if there is one.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4165 c0722133-6baf-4dd8-8699-98d999cd4f06