Progress on Apache2 compatibility -- doesn't work yet.
I've gotten to the point where the login screen can be displayed, but
login will still fail. The issue that i'm currently working on is that
we modify $r->param values extensively in WeBWorK, and that is not
supported in Apache2::Request. I'm not sure why, and no workaround is
given. One solution is to wrap override param() in WeBWorK::Request, but
this would slow things down since the Apache2::Request param method is
written in C.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4192 c0722133-6baf-4dd8-8699-98d999cd4f06