These two files can be used to render a .pg problem from the command line using the WebworkWebservice.
They are designed to be used as Unix Filters with BBedit, but can be tweaked to work
on any unix system and probably on windows as well.

These can be useful when you are writing many problems and wish to use your favorite
editor, and don't have direct access to the unix server.

The _rawoutput  version prints and HTML image of the raw XML data returned from the
webservice.  The other returns only the HTML page embedded in the XML data.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4287 c0722133-6baf-4dd8-8699-98d999cd4f06