This file helps WW manage the browsers windows that it opens.  The current approach is to use a separate window for the Problem Editor and a third window for viewing the problem when you test it.  This is fine, except that if you use the navigation links in the editor, those pages will open int he editor window, and this can mean the editor window is used for non-editing functions.  Worse, if you click on an "Edit it" link in this window and expect a new one to open, you will be surprised to see it appear in the current window.
This javascript file makes the editing window work as it was intended:
clicking on the navigation links opens those pages in the ORIGINAL
window again, not the editor window, so the editor window will be used
only for editing.  It also uses the main window for viewing the
problems, so there are only two main windows now.  This should
simplify the editing of problems.

To use this file, add

  <script src="<!--#url type="webwork" name="htdocs"-->/js/ww-windows.js"></script>

to the .template file that you are using just before the </body> (not
that it must be at the END of the page, or it may not work properly).

This will also fix a problem with Firefox where the editor window
woudl not come to the front if it already exists and inot the
front-most window.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4391 c0722133-6baf-4dd8-8699-98d999cd4f06