Added class to protect objects from evil prying problem code.
Uses the closure-as-object technique to protect the self hash of the
underlying object from problem code. All that the problem code sees is a
blessed coderef. The coderef is a closure over the underlying object and
the list of allowed methods. You can invoke the coderef directly, with
the name of the method to invoke as the first argument, but the class
also provides an AUTOLOAD routine that calls the coderef for you, so you
can continue to use the class as usual.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4395 c0722133-6baf-4dd8-8699-98d999cd4f06