removing courses directory (please read)
i originally moved the courses directory to courses.dist, to preserve
file history, but this wasn't the right strategy, since it caused the
courses directory to disappear from older branches as well. so i copied
it back to courses, and then used "cvs remove" to delete its contents.

so, it will hopefully exist in rel-2-2-dev and earlier releases, but not
in rel-2-3-dev or HEAD.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4441 c0722133-6baf-4dd8-8699-98d999cd4f06