Fix state retention when previewing or checking answers by retaining page but not worrying about preview or answer check.  Reduce PG translation for multipage tests by only translating problems on submission or as needed when checking.  Clean up some calls to get problems so that db->getAll... problems calls are used.  Add course logging for preview, newpage, and grading.  Add "Show Past Answers" form/button as appropriate.
Please note: there is an interesting question about the order in
which problems are logged for the case when problems on a test/quiz
are ordered randomly.  In this case, the problems are logged in the
course answer_log in the order in which they appear on the test.
However, when viewing grades and student progress, at the moment
this is not the case.  In the long run we probably want the grades
page to conform to the order shown in the test, while we may want
the student progress page to match the order seen in the actual
test definition.  For the moment that's probably a bug.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@4884 c0722133-6baf-4dd8-8699-98d999cd4f06