Update to version 3.4d of jsMath.
* Fixed a problem with easy/load.js where jsMath could fail to run when
easy/load.js is loaded at the top of the BODY of the document (rather than
the HEAD) and autoload is specified but tex2math is not needed (i.e.,
autoload is set to 1, while processSlashParens, processSlashBrackets,
processDoubleDollars, and processSingleDollars, are all set to 0 and
customDelimiters is commented out).

* Fixed a bug where easy/load.js could run its onload handler more than
once in some browsers.

* Fixed a bug in the autoload plugin where jsMath.js could be loaded twice
if a call to jsMath.Autoload.ReCheck() is made while the first copy of
jsMath.js is in the process of being downloaded.

* Fixed a subtle timing issue with jsMath.Process() when some mathematics
on the page requires an external file to be autoloaded and there are
additional jsMath.Process() calls pending.

* Modified the algorithm used by easy/load.js for finding the root when it
is not specified so that it will work better with MSIE in the situation
where easy/load.js is loaded via a relative URL.

* Added \operatorname to the AMSmath extension, which was acidentally left
out previously.

* Fixed some alignment problems with MSIE when running in "standards"
mode rather than "quirks" mode (i.e., in the presence of certain DOCTYPEs).


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@5456 c0722133-6baf-4dd8-8699-98d999cd4f06