Bring jsMath to version 3.4f:
* Fixed a problem with the \bbox extension where the sizes of boxed
items in super- and subscripts was incorrectly set, and where super-
and subscripts for large operators that are boxed might be placed
incorrectly.

* Version 3.4e arranged for tex2math to recognize
\begin{xxx}...\end{xxx} outside of math mode.  It incorrectly placed
the environment into in-line math mode rather than display mode.  This
had been corrected.

* One more attempt to work around alignment problems with MSIE with a
DOCTYPE that triggers "standards" mode.  I think I finally found the
real problem, and was able to remove several hacks that were not
completely effective.


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@5577 c0722133-6baf-4dd8-8699-98d999cd4f06