Added a banner which asks users to register production servers.
Clicking on the "register button" brings up a page with a mailto: link.
and a "server site registered" button.  Clicking on the mailto link brings
up the administrator's email application with suggested information to be emailed
to gage@math.rochester.edu (for now -- later we can replace this with an alias mail
address and eventually with some sort of form and a database for the recipient.)


Clicking on the "server site registered" button creates an empty file
courses/admin/registered_2.4.5

where 2.4.5 is the current version number found in WeBWorK.

if this file is present the "please register" banner goes away.

It's just a first cut at tackling this problem.

p


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@5664 c0722133-6baf-4dd8-8699-98d999cd4f06