Corrects bug (uncovered by newer versions of CGI.pm after version 3.29 and before 3.42) which defaulted students to "guests" in the permission popup menu used for editing the student record.   It showed up only for student because there default value was 0 -- which can be treated ambiguously -- as opposed to 2 or 10 which is not. Adding the override switch forces the default value generated in the perl code to override other default values for the popupmenu and provides the expected behavior.
The default value is set as [$default] so that [0] is properly recognized. This notation is usually used to select several
defaults however  just using $default did not work in CGI.pm 3.42.

This should fix the bug where students were mysteriously reassigned as guests and lost their "submit answer" buttons.

sites with older versions of CGI.pm will not have observed this problem.CVS: ----------------------------------------------------------------------


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@6186 c0722133-6baf-4dd8-8699-98d999cd4f06