modified stand alone client to provide courseName, user_id and password for identification these are in the file  webwork2/clients/.ww_credentials and webwork_xmlrpc_client.pl should be called from the client directory so that the credentials can be found.
The commands to listLib have been modified slightly.  The new commands are   'all', 'dirsOnly' and 'files'

WebworkWebservice.pm  now checks credentials for the user and password in the course specified.  At the moment
warning messages report the outcomes of authentication and authorization   but not use is made of this result.
and the behavior is as before. 


git-svn-id: http://svn.webwork.maa.org/system/trunk/webwork2@7076 c0722133-6baf-4dd8-8699-98d999cd4f06