Fix bug where scoredRecordedMessage is not defined.
This occurs around line 1222 in Problem.pm and appears to be the result of
mistakenly simplifying code so that values are not redefined.  This hot fix
has not been adequately tested -- and since I'm not sure why the previous
code was changed from the original in SVN -- this fix needs more investigation.