Break global.conf into  prelocal.conf, global.conf postlocal.conf
The entries in global.conf.dist will seldom be modified -- and in fact if
global.conf is not present global.conf.dist will be used instead.

The files prelocal.conf.dist and postlocal.conf.dist need to be copied to
prelocal.conf and postlocal.conf files and the local modifications applied.

database.conf.dist is seldom modified but needs to be copied to database.conf
in this version.