Make configurable which editors are shown in the left side bar.
This is a temporary? hack while we have code that has several alternative editors.  It allows
site managers to turn on or off one or more of the editors on a site wide or on
a course by course basis.  The basic configuration is in global.conf.dist  and there are
stubs for overrides in postlocal.conf.dist. One can add similar stubs to course.config.