Add error msg in xmlrpc when course not found.
During authentication verification, if the credentials specify a course
that does not exist, the DB object throws a TableMissing exception. This
exception ends up propagating to the xmlrpc handler as a generic
"Server: Application Error". So we just catch the exception and replace
it with a more informative "Course not found" error.

Also demonstrates how to transmit an error code with the error message
that was asked a few lines down in comments.