Implement xmlrpc createCourse, addUser, dropUser.
Same as the other xmlrpc functions, requires the same authentication
credentials to initiate a session. Also has the actual actions being
carried out by a separate module named CourseActions.

Beside the authentication credentials, each call needs these new entries
as part of the parameter.

createCourse:
  "name": "course name"
  "section": 100

addUser:
  "firstname": "John"
  "lastname": "Smith"
  "id": "The Doctor"         # username
  "email": "doctor@tardis"
  "studentid": 87492466
  "userpassword": "password" # defaults to studentid if empty
                             # if studentid also empty then no
                             # password
  "permission": "professor"  # valid values from %userRoles
                             # in global.conf
                             # defaults to student if empty
dropUser:
  "id": "The Doctor"         # username

Note that dropUser does not delete the user, only marks them as dropped
from the course. Because of this, addUser has to deal with two cases: an
actual new user or a re-enrolling user.

The method return is the same as the other xmlrpc calls, except with a
'status' entry marking the call as 'success' or 'failure'. 'failure'
means the action did not go through, there will be an accompanying
'message' entry with the error message.