Course creation webservice admin user conf.
During course creation using webservices, an admin user is added to the
course. This admin user is necessary since there isn't a classlist and
we need at least one user in the course. The admin user will be created
with the username and password specified in global.conf.dist.

These settings can be overwritten in postlocal.conf with:
$webservices->{courseActionsAdminUser} = "admin2";
$webservices->{courseActionsAdminPassword} = "blahblah";

The default admin user is "admin". It is recommended that the password
be changed from the default.