Limit webservice course creation to admin users.
Course creation webservice requests will only succeed if you are
successfully logged into the admin course and have instructor or above
permission.

This check is done in CourseActions. Note that there is a possibly
duplicate check of the permission level since the session initiation
function already checks for professor permissions.