Course creation webservice copies admin users.
On the regular webwork UI, when a new course is created, the instructors
from the admin course are copied over to the new course by default. This
allows admins easy access to the new course. The webservice course
creation call now also implements this by default.

The custom admin user configuration options are now dropped as a
result.

E.g., thse are no longer needed in postlocal.conf:
	$webservices->{courseActionsAdminUser} = "admin";
	$webservices->{courseActionsAdminPassword} = "admin";