Add use Carp; to file to prevent the program from dying when "croak" is called.
Minor fix.  There is an underlying issue, or possible issue, where there is a warning because
of a conflict between a login name and a cookie, possibly left over or stale.  Quitting the browser
fixes this. Is this a bug or a feature?  I suspect it's new behavior with the added LTI authentication.