Protect against syntax errors which kill the PG translation early so no pgcore object is produced.
I believe this was fixed earlier and somehow that fix got overwritten.  I'll go back
and look at where the overwrite occured when I have time.  What is in place
now is definitely a hack -- there are multiple copies of the error messages for example.
But at least it doesn't die outright with no useful error messages.