Fix bugs in MathTranslators.pm and SetActions.pm
Use a hash to pass parameters to CourseEnvironment->new rather than using
the deprecated method of passing four variables.  Using the deprated
method causes warnings to be reported in the debug.log (when debug is set
to 1) when apache is restarted.  This debug.log is then owned by root
and this causes many errors visible to students since apache can not
write to the debug.log file.