Be more careful about replacing _ with spaces in set names
Instead of replacing _ with a space a in $name itself we do the replacement in $display_name
The replacement is now only for the purposes of display and $name does not have spaces
which can sometimes cause problems.

this fixes a bug in which students (but only students) could not print pdf copies from the
Course homepage.