Fix bugs and improve error messages in SendMail.pm
It's a wonder that anything worked here.  It's clear that
errors in mailing would not have been reported.

The actual bug seems to be that the value $r needs to be passed
explicitly into cleanup_register().  This was not being done. It
is not clear why this used to work.

The error checking routine has been beefed up and to some extent
unit tested.  The unit tests can be repeated by uncommenting
the warn messages.