Replace publish by visible in database fields
This will cause some incompatibility for old courses.  It may require you
to reset whether a homework set is visible or not (once).  It will
also cause error messages in upgrade courses about extra fields in the database
which are not in the schema (e.g. published) -- these can be ignored.