Automatically create achievements directories.
If defaults.config file does not
have listings for the achievement directores then nothing happens.
All of these upgrade checks are skipped.

Check that modelCourses has the templates/achievements
and html/achievements directories.  If not then modelCourses
needs to be updated first -- issue a message to that effect.

If the course is missing either the templates/directory or the
html/achievements directory then create it from the corresponding
directory in modelCourse.