Added support for URLs as provided in WWfeedback messages
I had inadvertently provided an older version of CAS.pm when submitting
the original version.  This brings CAS.pm up to date.  The only change
was/is to add support to keep params on the URL being sent to the CAS server
(so that URLs provided by WWfeedback messages work properly).