Add drag and drop facility for installing languages.
Move a lang.po or lang.mo file to the Localize folder
(defined in defaults.config but set to lib/WeBWorK/Localize by default).

The name of the file will immediately appear in the General/Langauge
section of the Course Configuration page but it won't be possible
to switch to that language until an apache server restart has occurred.

Apparently mod_perl processes and stores the .po and .mo files so they
are not read on the fly.  Can this be changed?