Comment out check between key stored in cookie and key stored in request page
For now at least. There appear to be legitimate reasons
that the cookie key and the
url key don't agree in which case one should simply go with the page key.
These lines have popped annoying warnings in real usage cases.  I don't yet have
a catalog of all the times they occur.  You can provoke it by copying
a url with the key; logging out and logging back in again (which gives you
a new key) and then pasting the copied url back into the browser.

It has also been causing problems when WeBWorK is being used with Moodle.

Consult with Bill Wheeler on this.