Adjust appearance of homeworkset tables
Make text fit inside the text input boxes more neatly by using XXem as the unit of width -- set the width in math3.css
although it is also set inline for compatibility.  The font size is also reduced in math3.css for these tables.
Finally the labels for the checkboxes are removed and replaced by a link to the ProblemSetDetails page where the rest
of the data for the set can be edited.