Use files from local server for jquery and mathjax
this might result in using an earlier version of MathJax (at least for now) -- I don't know if that will cause bugs -- I haven't seen anything yet.
The links to jquery files will have to change when we move to the develop version where the js directory has been reorganized.
We will want to take the reorganization effort as an opportunity to centralize the delivery of jquery and other .js files so that we are not needlessly reloading these files.  If possible we should only need one version of jquery as well, not several versions.