Initial conversion to bootstrap layout system.
Nothing changed in the header for now.

The content area is now using bootstrap's fluid layout. There are 12 columns
total, the navigation takes up 2 columns while the body content takes up the
remaining 10 columns.

The footer is now in it's own distinct section.

Also changed the doctype to html5 and stripped out some of the old IE
workarounds that we shouldn't have to worry about anymore. At least, not
for math4, math3 will be the backward compatible theme.