Remove styles overriding bootstrap on the sidebar
The sidebar navigation was properly classed for bootstrap nav, but wasn't
working because some of the existing CSS was overriding it. Those were
removed.

The markup for the sidebar navigation was altered so that it's one long ul
element with nested ul childrens for bootstrap compatibility. Also replaced
the h2 "Main Menu" element with an li element that has the class nav-header.
And removed the strong tag from active elements.