Home page course listing & info-box css style.
Make the course list on the home page a list of easily clickable bars.

Removed fisheye as a selector. Note that since fisheye was set as an ID,
there are not supposed to be more than 1 element with that ID. Unfortunately,
it looks like there are times when more than 1 element with the fisheye
ID can be present on a page. This ruins the use of fisheye as a css selector
since some browsers will only apply the fisheye style to the first element
so named and ignore the rest.

Info-box was restyled to fit in with the new theme and to fit in properly
when they are generated as sidebar navigation additions.

Removed some old css that restyled elements, going to rely on the bootstrap
defaults for them.