Homework set & problem css changes.
Spacing issues with problem_set_body where it wasn't taking up as much
space as it should. Also found out what the functions of some of the elements
in system.template do, since they show up for the Homework Sets page. Took
the opportunity to rearrange and document them too.

I notice that the message section shows up twice, so I've decided to keep
the one on top and removed the duplicate at the bottom.