Fix Perl warning box not showing up.
I've somehow messed up the template so that the warning about Perl errors
doesn't show up anymore. Couldn't find what I did, so in the end, had to copy
and paste the known working version to fix it.

Took the opportunity to style the warning boxes using the bootstrap alert
boxes.