Banner shrink, add font-awesome for icons.
Shrunk the header images a bit to make for a smaller header.

Tried to use bootstrap navbar for the headers, but it turns out to be
unsuitable due to paddings they use to make text look nice. The paddings work
well for text, but not for images, the images are still too large and throws
evertyhing out of alignment.

Was planning on replacing the logout button with just the icon, but it turns
out to require changing perl code a bit, and I didn't want to deal with that.
Just added the icon to the button as a compromise for now.

Note that the WeBWorK logo was constructed with a transparent spider web image,
css, and text. It's not a whole image.

font-awesome overrides and expands on the default icons provided by bootstrap.
They're easier to style since they're not using bitmapped images but actual
fonts. The logout icon came from this set.