GatewayQuiz, remove generation of extra div & td.
There is an edge case where an empty gwMessage div is generated when there are
no recorded submissions & the test has expired. Fixed by moving the gwMessage div generation code into the if statements that prints the actual messages.

If there are multiple messages to show after checking a response, an extra td
is generated for every message after the first. This td would've been fine if
message rows don't generate the label td after the first message like how all
the other result rows are, but it generates the message label for every
message, so we end up with an extra blank td.