Add explanations for 'avoid_recording_answers' and 'don't_log_answers'.
The first gives permission to the user to NOT have their grades recorded.
The second means that those users and higher do not have their answers logged in the logfiles or in the pastAnswer database.  This can be used to keep TA and instructor answers from corrupting the data collected in those files.