Move hidden inputs for correct_ids and incorrect_ids to lie within the main problem form
Not sure if this is necessary -- but it does mean that these variables will now show up in listVariables()
It does not show in listVariables on the first submission. I am not sure why -- probably because they are calculated and
printed after the $inputs_ref hash is defined.  It shows on the second "submit button" push, I believe because
hidden fields are automaticaly copied over.

The fields correct_ids and incorrect_ids
are used by color.js which colors answer blanks  green or red depending on correctness.

Potentially this field can be used to determine whether or not a section in a compound problem can
be "opened".