Insure that display option radio buttons are listed in block form.
Geogebraweb calls  clean-css  from bootstrap which forces the radio buttons and labels into inline mode.
This makes the Display Options section ugly

Someone with more CSS knowledge might know of a cleaner way to do this.  What I've tried here
is to insist that in the math4 theme the labels inside the .optionView div are always blocks.
Since this is pretty specific selector it is likely that this css will be the one applied.