Stop updating problem_version and set_version tables for the admin course.
These are virtual tables and therefore not native.  The admin course updates non-native tables (ones that are global for all courses, not defined on a course by course basis) but these virtual tables do not need to be updated.

This is a hack.  If more virtual tables are created this fix will have to
be updated.

To check this you can use mysql commands to delete the version_id field from
problem_user table of the admin course.  Without this upgrade that version_id field will always be added back.  With this upgrade the version_id field, once deleted, will stay deleted from the database.

 Please enter the commit message for your changes. Lines starting