Fix the way WeBWorK rounds scores.  This fixes the bug reported in http://webwork.maa.org/moodle/mod/forum/discuss.php?d=3464
To test follow the procedure given on the forum. In short set up a
set with one 8 part problem (Library/272/setStewart14_5/problem_3.pg)
do one part correctlt( for a score of 12.5%). Check that all scores
are reorted consistently.  Also set up a set with the above
problem repeated 8 times and do 1/8 of each problem correctly.
Check that the total for the set is 1 even though each part is
reported to be .13