Change "alert" warnings to "console.log" warnings.
This means that flash applets placed in hints will emit warnings about the applet not being found (because it has not been inserted
into the page) but these warnings will be directed to the console.log and will not be observed unless one is using developer tools.

Check that this works on MSIE 8 and 9.  There were already lots of calls to console.log so things are not going to be worse.

This is not a perfect fix.  One wants to have a direct warning alert if an applet of some kind has not been loaded since otherwise one must infer indirectly the reason that the problem is not working as expected.  Having the warning delivered through the console.log is a
compromise.