Update documentation to show that "session_cookies" not "key" is the default value of $session_managment_via
This is a minimal fix for the moment.

The documentation about how the timestamp is maintained in session_cookies mode is not yet in alignment with the way the code works.  Before fixing that discrepancy (by changing the code and/or the documentation) we need to decide whether the cookie_timestamp or the database timestamp should take priority, or perhaps we should take the later one.

If one starts a session as an instructor for example and then starts a session as a student there are warning messages about conflicts between the cookie data and the hidden variable data embedded in the html page.  It seems to me the latter should take priority -- at least in this  case.  Comments?