Check explicitly that the database copy of the password is not blank or null.
Moodle wwassignment stores null passwords with the intent that access for that user cannot be obtained with any password, only with a session key.
This check explicitly checks for null or blank passwords (ones that don't match /\S/ ) instead of relying on side effects of the crypt function.

Minor change -- make sure $user_requests_cookie is always defined in order to avoid warnings.