Update the explanation in defaults.config explaining the current status of $session_managment_via="session_cookie".
Most of the functionality (e.g. using the cookie for a timeout clock) is disabled as a potential security hole.  The
"session_cookie" mode still automatically sets a cookie with a session_key which can be used if you loose the
session_key contained in the hidden variable on the html page.